+++ src/SqlArtisan/Internal/SqlPart/Keywords.cs
@@ -7,6 +7,7 @@ internal static class Keywords
     internal const string AddMonths = "ADD_MONTHS";
     internal const string All = "ALL";
     internal const string And = "AND";
+    internal const string Any = "ANY";
     internal const string Apply = "APPLY";
     internal const string As = "AS";
     internal const string Asc = "ASC";
@@ -142,6 +143,7 @@ internal static class Keywords
     internal const string Set = "SET";
     internal const string Sign = "SIGN";
     internal const string Skip = "SKIP";
+    internal const string Some = "SOME";
     internal const string Source = "SOURCE";
     internal const string Sqrt = "SQRT";
     internal const string StringAgg = "STRING_AGG";